+++ heligym/envs/dynamics/lookup.py
@@ -88,7 +88,7 @@ class LookUpTable:
         else:
             self._rowCounter = 0
             self._colCounter = 1
-        self._data = np.zeros((self._nRows + 1, self._nCols + 1), dtype=np.float)
+        self._data = np.zeros((self._nRows + 1, self._nCols + 1), dtype=np.float32)
         self._lastRowIndex = self._lastColIndex = 2
     
     def __lshift__(self, n):
@@ -96,7 +96,7 @@ class LookUpTable:
             Filling the table with `<<` operator.
         """
         assert self._rowCounter <= self._nRows, "There is no space in table to fill. Check the table size!"
-        self._data[self._rowCounter][self._colCounter] = np.float(n)
+        self._data[self._rowCounter][self._colCounter] = np.float32(n)
 
         if (self._colCounter >= self._nCols):
             self._colCounter = 0
